+++ wikidata/client.py
@@ -9,7 +9,7 @@ import urllib.parse
 import urllib.request
 import weakref
 
-from .entity import Entity, EntityId
+from .entity import Entity, EntityId, EntityType
 
 __all__ = 'WIKIDATA_BASE_URL', 'Client'
 
@@ -26,12 +26,28 @@ class Client:
     :param opener: The opener for :mod:`urllib.request`.
                    If omitted or :const:`None` the default opener is used.
     :type opener: :class:`urllib.request.OpenerDirector`
+    :param entity_type_guess: Whether to guess :attr:`~.entity.Entity.type`
+                              of :class:`~.entity.Entity` from its
+                              :attr:`~.entity.Entity.id` for less HTTP
+                              requests.  :const:`True` by default.
+    :type entity_type_guess: :class:`bool`
+
+    .. versionadded:: 0.2.0
+       The ``entity_type_guess`` option.
 
     """
+
+    #: (:class:`bool`) Whether to guess :attr:`~.entity.Entity.type`
+    #: of :class:`~.entity.Entity` from its :attr:`~.entity.Entity.id`
+    #: for less HTTP requests.
+    #:
+    #: .. versionadded:: 0.2.0
+    entity_type_guess = True
 
     def __init__(self,
                  base_url: str=WIKIDATA_BASE_URL,
                  opener: Optional[urllib.request.OpenerDirector]=None,
+                 entity_type_guess: bool=True,
                  repr_string: Optional[str]=None) -> None:
         if opener is None:
             if urllib.request._opener is None:  # type: ignore
@@ -43,18 +59,43 @@ class Client:
         assert isinstance(opener, urllib.request.OpenerDirector)
         self.base_url = base_url
         self.opener = opener  # type: urllib.request.OpenerDirector
+        self.entity_type_guess = entity_type_guess
         self.identity_map = cast(MutableMapping[EntityId, Entity],
                                  weakref.WeakValueDictionary())
         self.repr_string = repr_string
 
-    def get(self, id: EntityId) -> Entity:
-        """Get a Wikidata entity by its ID."""
+    def get(self, entity_id: EntityId) -> Entity:
+        """Get a Wikidata entity by its :class:`~.entity.EntityId`."""
         try:
-            entity = self.identity_map[id]
+            entity = self.identity_map[entity_id]
         except KeyError:
-            entity = Entity(id, self)
-            self.identity_map[id] = entity
+            entity = Entity(entity_id, self)
+            self.identity_map[entity_id] = entity
         return entity
+
+    def guess_entity_type(self, entity_id: EntityId) -> Optional[EntityType]:
+        """Guess :class:`~.entity.EntityType` from the given
+        :class:`~.entity.EntityId`.  It could return :const:`None` when it
+        fails to guess.
+
+        .. note::
+
+           It always fails to guess when :attr:`entity_type_guess`
+           is configued to :const:`False`.
+
+        :return: The guessed :class:`~.entity.EntityId`, or :const:`None`
+                 if it fails to guess.
+        :rtype: :class:`~typing.Optional`\ [:class:`~.entity.EntityType`]
+
+        .. versionadded:: 0.2.0
+
+        """
+        if not self.entity_type_guess:
+            return
+        if entity_id[0] == 'Q':
+            return EntityType.item
+        elif entity_id[0] == 'P':
+            return EntityType.property
 
     def request(self, id: EntityId) -> Union[
         bool, int, float, str,