+++ CHANGES.rst
@@ -8,6 +8,9 @@ To be released.
 
 - Added :attr:`Entity.type <wikidata.entity.Entity.type>` property and
   :class:`~wikidata.entity.EntityType` enum class to represent it.
+- Added :attr:`~wikidata.client.Client.entity_type_guess` option and
+  :meth:`~wikidata.client.Client.guess_entity_type()` method to
+  :class:`~wikidata.client.Client` class.
 
 
 Version 0.1.0